+++ CHANGELOG.md
@@ -29,6 +29,9 @@
 
 ### Added
 
+- [#135] (https://github.com/ChainSafe/forest-explorer/pull/135) Added link to
+  request for faucet top-up which is configurable using github env var.
+
 ### Changed
 
 ### Removed